docs: fix stale/inaccurate README; restore (and verify) the gcc claim
README had drifted from today's changes: the startup-init-order
paragraph described deleted functions (init_zobrist/init_eval/
init_search) — replaced with the no-globals Engine architecture; the
move-generation section still described generate_legal as pseudo+filter
under the old is_legal_fast name — now the masked single-pass design
with current names and the ~550 Mnps figure; the evaluation seam
signature gained its EvalCache parameter and the finny cache is
per-Searcher (not thread-local) with Engine-owned nets; the move-
ordering bullet wrongly claimed SEE orders captures (it prunes them);
the Verify section now states what CI actually runs cross-platform.

The "gcc 13+ works" claim had become false: gcc rejects comparing
distinct anonymous enums in the bucket static assert (cast to int) and
does not implement #pragma STDC FP_CONTRACT (now clang-only, with the
one contractible expression in portable_log split across statements so
neither compiler can fuse it under -std=c17). gcc 13.3 now builds
-Werror-clean and produces the identical bench signature 2657379 and
perft counts — the README claim is upgraded to that stronger, verified
statement.

+++ src/nnue.c
@@ -39,7 +39,7 @@ struct NnueNetwork
     int32_t output_bias;                    ///< output bias
 };
 
-_Static_assert(NNUE_KING_BUCKETS == NUM_KING_BUCKETS, "header/implementation bucket counts must agree");
+_Static_assert((int)NNUE_KING_BUCKETS == (int)NUM_KING_BUCKETS, "header/implementation bucket counts must agree");
 
 /**
  * @brief King-input bucket for a perspective-relative king square: 4 file-pairs x 2 board-halves.
